0.0.2

+++ src/main.rs
@@ -3,7 +3,7 @@ fn main() {
     std::io::stdin()
         .read_line(&mut user_input)
         .expect("Failed to read user input!");
-    let num_1: u8 = user_input
+    let num_1: u128 = user_input
         .trim()
         .parse()
         .expect("Failed to parse user input!");
@@ -13,14 +13,14 @@ fn main() {
     std::io::stdin()
         .read_line(&mut user_input)
         .expect("Failed to read user input!");
-    let num_2: u8 = user_input
+    let num_2: u128 = user_input
         .trim()
         .parse()
         .expect("Failed to parse user input!");
     let gcd_vec_1 = factor(num_1);
     let gcd_vec_2 = factor(num_2);
 
-    let mut final_gcd: u8 = 0;
+    let mut final_gcd: u128 = 0;
     for i in &gcd_vec_1 {
         if gcd_vec_2.contains(i) {
             final_gcd = *i;
@@ -29,7 +29,7 @@ fn main() {
 
     println!("gcd: {final_gcd}");
 
-    let mut multiplier: u8 = 1;
+    let mut multiplier: u128 = 1;
     loop {
         let multiplied = num_1 * multiplier;
         if multiplied % num_1 == 0 && multiplied % num_2 == 0 {
@@ -41,7 +41,7 @@ fn main() {
     }
 }
 
-fn factor(number: u8) -> Vec<u8> {
+fn factor(number: u128) -> Vec<u8> {
     let mut factor_vec: Vec<u8> = Vec::new();
     for i in 1..number {
         if number % i == 0 {
